+++ cfbd_json_py/teams.py
@@ -2,8 +2,7 @@ def get_cfbd_team_information(
         api_key: str = None,
         api_key_dir: str = None,
         conference_abv: str = None,
-        cache_data: bool = False,
-        cache_dir: str = None,
+
         return_as_dict: bool = False):
     """
 
@@ -17,8 +16,7 @@ def get_cfbd_fbs_team_list(
         api_key: str = None,
         api_key_dir: str = None,
         year: int = None,
-        cache_data: bool = False,
-        cache_dir: str = None,
+
         return_as_dict: bool = False):
     """
 
@@ -33,8 +31,7 @@ def get_cfbd_team_rosters(
         api_key_dir: str = None,
         team: str = None,
         year: int = None,
-        cache_data: bool = False,
-        cache_dir: str = None,
+
         return_as_dict: bool = False):
     """
 
@@ -48,8 +45,7 @@ def get_cfbd_team_talent_rankings(
         api_key: str = None,
         api_key_dir: str = None,
         year: int = None,
-        cache_data: bool = False,
-        cache_dir: str = None,
+
         return_as_dict: bool = False):
     """
 
@@ -66,8 +62,7 @@ def get_cfbd_team_matchup_history(
         api_key_dir: str = None,
         min_year: int = None,
         max_year: int = None,
-        cache_data: bool = False,
-        cache_dir: str = None,
+
         return_as_dict: bool = False):
     """