+++ cfbd_json_py/venues.py
@@ -3,8 +3,6 @@ def get_cfbd_venues(
         api_key: str = None,
         api_key_dir: str = None,
 
-        cache_data: bool = False,
-        cache_dir: str = None,
         return_as_dict: bool = False):
     """